remove RevisionHash and change all record helper methods to accept raw HeaderHash in preparation for indexing logic changes
DnaAddressable now must be directly storable as an Entry in order for fully-qualified IDs to be encoded in the DHT. This means conversion traits to/from hdk::prelude::Entry must be created, and the inner type of DnaAddressable cannot be ambiguous since this would make implementing such conversions impossible. This makes sense and is a good thing, because it enforces that only tuples of (DnaHash, EntryHash) can be used as identifiers. Implementors should not be linking to revisions but rather records.